Move DisplayedTask into task_editor.rs
